+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Invalidate access token cookie on cas redirect ([#8](https://github.com/scm-manager/scm-cas-plugin/pull/8))
+
 ## 2.1.0 - 2020-09-01
 ### Added
 - Integrate full anonymous access mode ([#5](https://github.com/scm-manager/scm-cas-plugin/pull/5))